split db from index tests
